More Bazel setup and dealing with external dependencies.
This commit extends the system we did last time, but now adds the
ability for the shared library and the days themselves to use crates.

There is a lot of Bazel magic, but it is fairly straightforward. The
main module loads an extension, which in turn creates "rules" for
building "all_crate_deps" which can then be used as one of the
dependencies in the individual BUILDs. This results in crates being
centrally managed by the build system itself. The first run is dreadful
slow, but that is just par for the course in Rust.