Add the second part of the secon day.
This was really satisfying to solve. The ability to create all
combinations of length using itertools was a godsend and allowed me to
focus on the main logic quite quickly.

There are a few refactorings I would probably do if this were production
code, mainly try to understand why the references get wonky with
`itertools::Combinations` (see the ugly vector transform in Part Two)
and using a PartitionMap to generate the "safe by default" and
"potentially safe" sets.