Add the second part of day 03.
Still easy, required named capture groups in order to make the API nice.
This is a good example where I feel like there should be a pure
functional way to solve this, but my knowledge of arcane itertools/std
functions fails me.

The part I really don't like is how we must yield an element in the
capture group map, so that we can sum it in the end. But in the
imperative version, we would have a very clear branch. I don't know what
I want it to look like really, but I have a feeling this ain't it.