Add the solution for the first part of day 04.
This was just annoying more than anything. I didn't have the brain power
to actually write out a clever algorithm for this, so I Googled some
solutions and made them more readable.

However, this is going to be the target of my SIMD optimisation for
sure, as I am sure that it will just be more fun to try to do this with
fancy instructions and not straight loops. (the iterators and functional
style really really help here!)